Update epsilon type hint from optional int to optional float in all rms norm config

+++ examples/modular-transformers/configuration_duplicated_method.py
@@ -122,7 +122,7 @@ class DuplicatedMethodConfig(PreTrainedConfig):
         hidden_act: Optional[str] = "silu",
         max_position_embeddings: Optional[int] = 2048,
         initializer_range: Optional[float] = 0.02,
-        rms_norm_eps: Optional[int] = 1e-6,
+        rms_norm_eps: Optional[float] = 1e-6,
         use_cache: Optional[bool] = True,
         pad_token_id: Optional[int] = None,
         bos_token_id: Optional[int] = 1,
